requirements.txt
asgiref==3.5.0
Django==4.0.3
django-cleanup==6.0.0
django-jazzmin==2.4.9
django-rest-framework==0.1.0
djangorestframework==3.13.1
gunicorn==20.1.0
Pillow==9.0.1
psycopg2==2.9.3
pytz==2021.3
sqlparse==0.4.2
tzdata==2021.5
whitenoise==6.0.0